+++ tuneful-client/src/UserProfile/EditProfile.js
@@ -7,9 +7,9 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import { makeStyles } from '@material-ui/core/styles';
-import theme from '../../theme/instapaper/theme';
+import theme from './instapaper/theme/instapaper/theme';
 import Slide from '@material-ui/core/Slide';
-import useForm from "../../../../CustomHooks/useForm";
+import useForm from "../CustomHooks/useForm";
 
 const Transition = React.forwardRef(function Transition(props, ref) {
     return <Slide direction="up" ref={ref} {...props} />;
